+++ maintenance-bungee/src/main/java/eu/kennytv/maintenance/bungee/MaintenanceBungeePlugin.java
@@ -33,6 +33,8 @@ import eu.kennytv.maintenance.core.dump.PluginDump;
 import eu.kennytv.maintenance.core.hook.ServerListPlusHook;
 import eu.kennytv.maintenance.core.proxy.MaintenanceProxyPlugin;
 import eu.kennytv.maintenance.core.proxy.SettingsProxy;
+import eu.kennytv.maintenance.core.proxy.util.ProfileLookup;
+import eu.kennytv.maintenance.core.proxy.util.ProxyOfflineSenderInfo;
 import eu.kennytv.maintenance.core.util.SenderInfo;
 import eu.kennytv.maintenance.core.util.ServerType;
 import eu.kennytv.maintenance.core.util.Task;
@@ -180,7 +182,18 @@ public final class MaintenanceBungeePlugin extends MaintenanceProxyPlugin {
     @Nullable
     public SenderInfo getOfflinePlayer(final String name) {
         final ProxiedPlayer player = getProxy().getPlayer(name);
-        return player != null ? new BungeeSenderInfo(player) : null;
+        if (player != null) {
+            return new BungeeSenderInfo(player);
+        }
+
+        final ProfileLookup profile;
+        try {
+            profile = doUUIDLookup(name);
+        } catch (final IOException e) {
+            e.printStackTrace();
+            return null;
+        }
+        return new ProxyOfflineSenderInfo(profile.getUuid(), profile.getName());
     }
 
     @Override