+++ .github/CHANGELOG.md
@@ -4,6 +4,7 @@ This file contains update logs for this project. The top may contain a `Unreleas
 ---
 ## Unreleased
 ### Changed
+* Bungee/Velocity: `/maintenance add <name>` now works for offline players as well
 * Added config options `commands-on-maintenance-enable` and `commands-on-maintenance-disable` to define commands to be executed after maintenance has been toggled
 * Added config option `enable-pingmessages` to can change whether a custom maintenance motd should be displayed during maintenance
 * Added language fields `scheduletimerBroadcast` and `singleScheduletimerBroadcast` with the variables `%SERVER%`, `%TIME%`, and `%DURATION%` instead of using the normal starttimer broadcasts